+++ Changes.md
@@ -19,7 +19,12 @@ Breaking Changes
 - ArnoldShader : The `standard_volume` shader is now assigned via an `ai:volume` attribute instead of `ai:surface`.
 - RenderUI : Removed deprecated `rendererPresetNames()` function.
 
-1.6.x.x (relative to 1.6.5.1)
+1.6.x.x (relative to 1.6.6.0)
+=======
+
+
+
+1.6.6.0 (relative to 1.6.5.1)
 =======
 
 Features
@@ -36,9 +41,15 @@ Improvements
   - Added Interpretation row for primitive variables, with values of "Point", "Vector", "Normal" or "UV".
   - Added editing capabilities to the Local Transform properties.
 - LightEditor : Added Transform section, with columns displaying the light's transform.
+- RenderPassEditor :
+  - Added "Favourites" section. Columns from any other section can be marked for display in this section by right-clicking on their header and selecting "Favourite" from the menu. Columns can be reordered within the Favourites section by dragging their header.
+  - Any standard or renderer-specific option can be added as a favourite by clicking on the `+` column header and choosing it from the menu, or any option can be added via drag & drop from the SceneInspector.
+  - A default set of favourite columns can be configured by registering metadata containing a list of option names in a startup file. `Gaffer.Metadata.registerValue( GafferSceneUI.RenderPassEditor.Settings, "favouriteColumns", "userDefault", IECore.StringVectorData( [ "option:render:cameraInclusions", "option:render:matteInclusions" ] ) )`
 - ImageToTensor : Added `tensorElementType` plug to choose between `Float`, `Float16` and `BFloat16`.
 - TensorToImage : Added conversion from `Float16` and `BFloat16` tensor elements in addition to the existing `Float` support.
 - RenderManOptions : Added XPU device configuration options.
+- `gaffer dispatch` : Removed restrictions that prevented nodes such as Switch and ContextVariables from being dispatched.
+- Spreadsheet : Improved visual feedback when reordering columns via drag & drop.
 
 Fixes
 -----
@@ -46,11 +57,13 @@ Fixes
 - SceneInspector : Fixed sporadic incomplete updates, particularly when an interactive render was running.
 - Render, InteractiveRender, StandardOptions : Fixed bugs allowing Cycles and 3Delight to appear as available renderers even when hidden from the UI or not configured.
 - Options, Attributes : Fixed missing presets.
+- CyclesShader : Fixed unstable ordering of parameter presets. These are now sorted alphabetically.
 
 API
 ---
 
 - GridContainer : Added `nextRow()` context manager, to simplify the process of arranging child widgets in rows.
+- SceneInspector : Added `draggedAttributes()`, `draggedOptions()` and `draggedParameters()` static methods.
 
 1.6.5.1 (relative to 1.6.5.0)
 =======